Removed "convertLFtoCR" HL7v2 property from UI, references outside the ImportConverter, and non-deprecated classes. Added inputSegmentDelimiter and outputSegmentDelimiter to HL7v2 properties. This replaces the convertLFtoCR functionality. The default settings for these two should behave identical to convertLFtoCR being true. To mimic convertLFtoCR being false, set the inputSegmentDelimiter to "\r". Any other settings will cause new behavior that is not possible in 2.x. Through this, it is now possible to use segment delimiters other than "\r" in the encoded content.
Added isTransformerRequired() method to XmlSerializer interface. This replaces the optimization that checked if data type properties were equal to the default. The current implementations provide the same result, but allows for flexibility in the future to allow non-default properties to NOT serialize/transform the message.
Added transformWithoutSerializing() method to XmlSerializer interface. This allows transformations to occur without serialization to xml first. It will only ever be called if inbound and outbound data types are the same. 

Removed some methods from IXMLSerializer that are no longer used.

Removed DefaultSerializerPropertiesFactory and moved its method to SerializerFactory. There's no need to have both.
Issue: MIRTH-2272