Fixes for database dispatcher
Don't replace template values for a non-javascript query in getReplacedConnectorProperties so it can be replaced later.
Fixed the javascript executor using the wrong scope that did not include the messageObject.
Fixed error messages not being saved.
Fixed the data source not being reopened after a channel restart.
Fixed the old data source not being closed properly if the properly a new one is created because the properties have changed.
Added code to reattach attachments (only works for SQL right now, not for javascript).

Allow null character encoding to be passed to reAttachMessage to indicate we just want to use the default attachment encoding (UTF-8)
Added method to determine whether a string has attachment keys in it.
Issue: MIRTH-2250