If source or destination queues are enabled and the current channel storage settings are not compatible with queueing, throw a deploy exception.
Updated the channel setup ui so that an error message is shown if the user attempts to enable the source queue and the storage setting does not permit using a source queue.

Also fixed a problem when a deploy exception occurs, the channel remained in Donkey's list of deployed channels. Donkey now removes the channel from the list if a DeployException occurs so the channel will not appear in the dashboard.
Issue: MIRTH-2506