Fixed custom metadata search not showing the correct operators for the metadata type.
No longer setting the value in the custom metadata search table after the value fails to cast to the metadata type.

MirthComboBoxTableCellEditor now always fires editing stopped, regardless of whether a custom action listener was passed in.

Renamed the private ComboBoxCellEditor in DataTypesDialog to correctly be called CheckBoxCellEditor.
Issue: MIRTH-2521