MIRTH-2552: Updated VmDispatcherProperties migration - added responseTimeout field.
Replaced the DonkeyElement interface with the implementation that was previously defined in MirthElement. Removed the MirthElement class.

Updated ObjectXMLSerializer.fromXML(source, expectedClass) so that it uses generics and now returns the deserialized object of the appropriate type.

Added ObjectXMLSerializer.listFromXML(), which returns a deserialized list object when a list may be expected.

Updated ImportConverter3_0_0.convert() to be able to migrate a list of alerts.

Updated 3.0.0 connector migration so that it reads the connector type from the 'transportName' instead of the 'DataType', which is not guaranteed to be present in all connector exports.

Updated serialization logic so that the 'version' attribute is only set on the top-level document element and not on any child elements of the serialized document, since it is only needed on the document element.

Removed the 'version' field from the Channel class since it is not needed for any migration code.


MIRTH-2599:
Added migration code for database receiver/dispatcher properties.