Brought back the AlertSender class and reference list item. AlertSender now dispatches events instead of calling the alert controller directly. AlertSender is now constructed with either channelId or a connector message. The connector message allows it to dispatch error events for a specific connector.
Events dispatched from the AlertSender that do not have a metaDataId associated will trigger if any of the connectors for that channel are enabled.
Events dispatched from the AlertSender will only trigger if the User Defined Transformer event type is checked.
User javascript code for sending alerts should remain identical to how it was in 2.x

Cleaned up the error message builder to only append new lines when needed.
Issue: MIRTH-2621