MIRTH-2471 MIRTH-1699
Changed Derby's donkey schema to allow cascade deletion on foreign keys.

Updated Derby's pruner statements to work with the cascade deletion and support pruning based on d_mm.received_date instead of d_m.received_date.

Refactored DataPruner so that it can delete messages based on a list of message ids even when not archiving. If a pruning strategy has been explicitly defined or if the current database has the "pruneMessagesByIds" query defined, it will automatically prune based on a message id list, rather than a date threshold. This improves performance dramatically in Derby.

Cleaned up code in DataPruner and made it easier to read.

Removed the MessageExportResult class since it is no longer needed.

Cleaned up and fixed problems with the DataPrunerTest unit tests.