Fixed outdated channel being deployed with no warning in certain cases if a channel is re-imported and deployed immediately after it is deleted.
Modifed updateChannel to use getCachedChannelById instead of getChannel to determine whether a channel already exists. This implicitly updates the channel cache before a channel is saved.

Synchronized updateChannel and removeChannel in DefaultChannelController to ensure that the channel cache and database can never have different versions of a channel at the same time.
Issue: MIRTH-2642