Fixed bug with migrating unrecognized connectors, the ExtensionController was being used, which wouldn't work on the client side.
Hardcoded a reference to a migrator class that will run migration on the Email Reader connector if it is detected.

Added Migrator.getUninstallStatements() method which will allow plugins to return a list of sql statements to uninstall the plugin from the database.

Extracted logic from Migrator.executeScript() that read statements from a sql file and put it into a readStatements() method so plugins can easily read statements from an uninstall script file to return via getUninstallStatements().

Added Migrator.tableExists() so that plugin Migrator classes can detect if a table exists. Replaced similar logic in ServerMigrator with a call to this method.

Added a new abstract method to Migrator: migrateSerializedData() - in case plugins need to migrate serialized content in the database as ServerMigrator does.
Issue: MIRTH-2647