MIRTH-2804 Removed unused indexes.
Determined that Postgres does not use the index on (message_id, id) even when those are the only columns in the where clause. It uses the (messageid, id, status) index instead.

MIRTH-2792
Added new index to the d_m table to account for queries that now must scan for both id and server_id.