JmsDispatcher is now thread safe and supports multiple queue threads.
JmsDispatcher now creates connections lazily and caches them for future use. Connections are shared between threads, but sessions and producers are not. If a message fails to send and a cached connection was used, it will attempt to recreate the connection once and send the message again.

JmsDispatcherProperties no longer contains a reconnect interval. The reconnect interval in JmsConnectorProperties has been moved to JmsReceiverProperties.


Issue: MIRTH-2791