MIRTH-3037
Added the ability for attachment handlers to override the re-attachment logic that is run immediately before a message is sent out from a destination connector.

Removed the MessageAttachmentUtil class and moved it's methods into a new MirthAttachmentHandler class, which is now the super class of the built-in attachment handlers. This way, attachment handlers can now override and extend the re-attachment logic.

Updated destination connectors and other places that previously called MessageAttachmentUtil to instead retrieve the attachment handler from the deployed channel and call the appropriate reAttach method.

Converted the AttachmentHandler abstract class into an interface since all of it's methods were abstract.

Added a PassthruAttachmentHandler so that destination connectors don't have to do null checks if an attachment handler isn't defined.