MIRTH-3166: Improved how channels are enabled/disabled from the Channels view. Instead of doing a separate request for each channel (and retrieving the channels again after each request), it now only sends a single request via a new Client operation, setChannelEnabled. A set of channel IDs are passed in along with a boolean that determines whether to enable/disable.
On the server side in the channel controller, setChannelEnabled is synchronized with update/remove so that it can guarantee the state of the cache. Then it iterates through each channel ID and updates the associated channel (if it exists and is valid). If one channel fails to be updated for any reason, the rest are still attempted.

On the client side, only a single SwingWorker is now used. It still checks beforehand whether some of the selected channels are invalid, but it no longer prevents the worker from executing in that case. Instead, after executing the worker it will alert the user at the very end if one or more channels could not be enabled. For disabling this isn't necessary since validation doesn't matter in that case.