Added a local channel ID column to the channel panel view.
A list of local channel IDs is cached from the server and stored in the default channel controller. The cached channel ID is then used to look up the local channel ID from this map variable, set to a local ID variable stored within the channel status.

This new column is accessed by right clicking the header row and selecting the option to show or hide the column. The visibility of this column is stored in the user preferences node to determine whether or not to display the column on startup.
Issue: MIRTH-3260