Added the ability to use velocity template variables when routing via a channel writer.
The new channel ID field will display the unique guid for any channel selected using the combo box. Alternatively, the combo box now dynamically updates based off of what's typed into the channel ID field. If the text matches a channel's unique id, the combo box will display the name of the channel, however, if a channel does not exist for the entered text, the combo box will default to "<Channel Not Found>". Also, if a velocity template variable is used or is empty, the combo box will reflect this state as well with a corresponding option, "<Map Variable>" or "<None>".

This feature is achieved by adding a document listener to the underlying field to handle the aforementioned logic. Because these two fields are now coupled, additional logic has been placed to ensure that an infinite loop does not occur when the user types into the field.
Issue: MIRTH-2961