Refactored the package and variable names for Channels and Connectors in DonkeyEngineController. Since we have Channel and Connector in two different packages (model and donkey), we have always referenced the donkey channel and connectors fully qualified. However since we are in DonkeyEngineController, it makes more sense to reference the Donkey channels directly and the model channels fully qualified.
We always refer to the donkey channel/connector as channel and connector. These are the actual channel and connectors that are deployed. We always reference the model channel and connectors as the "model". Therefore we've renamed all references to the donkey channels as just "channel" and all references to the model channels as "channelModel" which is more in line with how we normally refer them. Likewise for "connector" and "connectorModel".

Renamed convertToDonkeyChannel to createChannelFromModel to better describe what is happening. We aren't actually converting anything since both the channel and model still exist.
Issue: MIRTH-3393