Reverting to DBCP 1.4 since the snapshot build had a new bug.
Created the class PoolableConnection in the dbcp package so we can "override" the jar without recompiling it.

Backporting the fix for https://issues.apache.org/jira/browse/DBCP-391 into the 1.4 release.
This fix actually introduced a new bug that was only fixed on the DBCP2 branch at revision 1568674.
Therefore I've also backported the fix for the new bug as well. It will (probably) work.
If we ever update the DBCP jar or switch to a different connection pool then this class will need to be removed.
Issue: MIRTH-3262