MIRTH-3351
Added the ability for extensions to override any of the controller classes represented in ControllerFactory, as well as the DataPrunerController, by putting a <controllerClasses> element in plugin.xml.

Moved logic that reads the available extensions from the filesystem from DefaultExtensionController into a separate 'ExtensionLoader' class that has no dependencies on any controller classes.

Removed hardcoded references to controller classes from specific extensions in DefaultControllerFactory, DataPrunerController and (donkey) ChannelController.

Added wrapper methods in ChannelController to Donkey's ChannelController to get and reset statistics, so that these methods can be overridden by an extension.