The BasePanel getData implementation now cycles through the card layout to retrieve the appropriate script data based on the visible card.
An abstract method to get the generated script was added to the MirthEditorPanePlugin so that each individual rule/step can provide its own implementation, and several scripts had their generated content formatted for readability.

Tabbed panes were added to both the filter and transformer panes so as to display the aforementioned scripts. Functionality was provided to update the code tab when appropriate.
Issue: MIRTH-2128