A new block size field is available on the Data Pruner settings panel, with a default size of  100 unless otherwise specified by the user.
The validation method within the MessageExportPanel was refactored to return a string instead of a boolean value. This was done to give specific details as to what errors might've occurred within the panel.

The tooltip for including attachments during export/archiving was also ninja hot fixed since it previously did not exist.
Issue: MIRTH-3531