MIRTH-3649: Fixed bug where creating database connections in JavaScript using custom library resources did not always work. DatabaseConnectionFactory will now attempt to create a Driver instance and use that as necessary. If custom libraries aren't being used, or they don't contain the driver class, then everything will default to DriverManager (how it was working before). Similar changes were made to the Database Reader/Writer. They will attempt to create a custom driver, and if that doesn't work, default to DriverManager.
What this means is that not only will DatabaseConnectionFactory work correctly in JavaScript, but also you can now use multiple versions of the same database driver in different channels/connectors. Before they would have conflicted because of the way DriverManager works, but CustomDriver eschews that nightmare of a class completely.