MIRTH-3167: Made tweak to the destination queue to more correctly handle rotation now that we allow multiple processing threads. Now that there are multiple simultaneous producers for the destination queue, it's possible that messages can be added to the buffer out of order (which is fine). This really only matters for rotation because we used the message IDs to determine when to sleep for the retry interval in the queue thread. Now instead of just checking the last message ID, we also go directly to the queue and check whether the current thread ID is flagged for rotation. This is done by the queue data source when the rotation actually occurs.
Because messages by multiple producers can be out of order, it's possible that a message with a higher ID can be committed to the database first. For queues with rotation on, this means that there can be bursts of smaller queries to grab messages _earlier_ than that first one. Overall this should be fine, since once those messages finish the max message ID will again be set and the next batch of messages will be retrieved.