MIRTH-3213 The MirthComboBoxTableCellEditor's underlying component can now be editable. If enabled, a custom editor is applied which is responsible for decorating the component with a border, allowing all the standard functionality of a simple editable field. If the field is editable, any free-text values are returned from the underlying editor instead of from the model directly.
The ActionTableModel will now directly set the free-text, when doing a lookup into the protocol options fails to return an entry. The local inner Protocol class has been modified to allow for these changes.

The ChannelProtocol may now accept recipients as either a name or an ID. The protocol will first attempt to retrieve the channel by ID and then by name if the above fails. If all else fails and a channel cannot be found, a warning is logged.
Issue: MIRTH-3213