MIRTH-3732: Overhauled tags and filtering on the Dashboard and Channels views.
Inside the Channel Edit view, the tags are now set using a single field component, rather than a table. On the Dashboard and Channels views, the wrench icon at the bottom-left has been replaced by the same field component. This filter field can be used to enter tags, specific channels, or free text. If free text is used, it will match on the channel name (ignoring case).

For the dashboard, the filtering is actually done on the server side, so it can potentially speed up dashboard refresh requests for instances that have a lot of channels.

On both views, there are new buttons in the bottom-right, to switch how tags are visible on each channel (either an icon with tool-tip, or a bubble-text label).

There is also a new settings tab for tag configuration. On this view the user can see at a glance how many tags there are across the entire server. The names and background colors can be changed, and channels can be bulk-selected across multiple tags at once. Tags can be added or removed from this view as well.

User Auth permissions for view/manage were added as well.